Move SAM1 training into micro_sam/v1/training submodule
