+++ setup.cfg
@@ -80,7 +80,7 @@ console_scripts =
     micro_sam.image_series_annotator = micro_sam.sam_annotator.image_series_annotator:main
     micro_sam.precompute_embeddings = micro_sam.precompute_state:main
     micro_sam.automatic_segmentation = micro_sam.v1.automatic_segmentation:main
-    micro_sam.train = micro_sam.training.training:main
+    micro_sam.train = micro_sam.v1.training.training:main
     micro_sam.evaluate = micro_sam.evaluation.evaluation:main
     micro_sam.info = micro_sam.util:micro_sam_info
     micro_sam.benchmark_sam = micro_sam.evaluation.benchmark_datasets:main
